+++ nue/model/__init__.py
@@ -1,0 +1,5 @@
+from .base import GPTConfig
+
+__all__ = [
+    "GPTConfig",
+]
