+++ nue/model/torch.py
@@ -11,23 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from dataclasses import dataclass
 from typing import cast
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-
-@dataclass(frozen=True)
-class GPTConfig:
-    vocab_size: int  # BPE vocab (same as GPT‑2)
-    ctx_len: int  # maximum sequence length
-    n_embed: int  # model dimension (d_model)
-    n_heads: int  # number of attention heads
-    n_layers: int  # transformer blocks
-    mlp_ratio: int  # feed‑forward expansion ratio
-    dropout: float = 0.0
+from .base import GPTConfig
 
 
 class RotaryEmbedding(nn.Module):
