+++ nue/train/__init__.py
@@ -1,9 +1,7 @@
 from .base import Epoch, TrainingOptions, TrainingSession
-from .trainer import PyTorchTrainer
 
 __all__ = [
     "Epoch",
-    "PyTorchTrainer",
     "TrainingOptions",
     "TrainingSession",
 ]
